New: JOSM reading, writing, merging changeset attribute fixed #4090: Add changeset:* search option to JOSM's search engine
git-svn-id: http://josm.openstreetmap.de/svn/trunk@2604 0c6e7542-c601-0410-84e7-c038aed88b3b

+++ src/org/openstreetmap/josm/io/OsmChangesetParser.java
@@ -71,9 +71,9 @@ public class OsmChangesetParser {
             if (value == null) {
                 throwException(tr("Missing mandatory attribute ''{0}''.", "id"));
             }
-            long id = 0;
+            int id = 0;
             try {
-                id = Long.parseLong(value);
+                id = Integer.parseInt(value);
             } catch(NumberFormatException e) {
                 throwException(tr("Illegal value for attribute ''{0}''. Got ''{1}''.", "id", value));
             }
